Add optional parameters to add comment action
This commit changes the 'add comment' action of the BMC Remedy app to
add two new optional action parameters.

1. description: This is a short description associated with the main
comment content (worklog). Some Remedy services are configured to accept
new comments only if the "Description" field is provided in the request.

2. worklog_submitter: Allows for submitting comments (worklogs) under
the username or email address of someone else. By default, the comments
will be submitted with the username or email address configured in the
asset settings.